Change to localhost

+++ log/src/helpers/SessionHelper.kt
@@ -1,8 +1,8 @@
 package com.jstarczewski.log.helpers
 
+import com.jstarczewski.log.db.UserDataSource
 import com.jstarczewski.log.LogSession
 import com.jstarczewski.log.db.User
-import com.jstarczewski.log.db.UserDataSource
 import io.ktor.application.ApplicationCall
 import io.ktor.application.call
 import io.ktor.sessions.get
@@ -10,6 +10,6 @@ import io.ktor.sessions.sessions
 import io.ktor.util.pipeline.PipelineContext
 
 suspend fun PipelineContext<Unit, ApplicationCall>.withSession(db: UserDataSource, block: suspend (User) -> Unit) =
-    call.sessions.get<LogSession>()?.let { db.userById(it.userId) }?.let {
+    call.sessions.get<LogSession>()?.let { db.userById(it.userId.toLong()) }?.let {
         block(it)
     }
